requirements.txt
fastapi
pydantic
aiosmtplib
python-dotenv
uvicorn
email-validator